LayoutTests/http/tests/security/contentSecurityPolicy/object-src-param-code-blocked-expected.txt

CONSOLE MESSAGE: Refused to load plugin data from 'http://127.0.0.1:8000/plugins/resources/mock-plugin.pl?code' because it violates the following Content Security Policy directive: "object-src http://localhost:8000".

This test passes if there is a console message saying the plugin was blocked. 
